% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aenetgt.R
\name{aic.bic.eric.enetgt.grid}
\alias{aic.bic.eric.enetgt.grid}
\title{Chooses tuning parameters via AIC, BIC, and ERIC.}
\usage{
aic.bic.eric.enetgt.grid(enetgt.grid.out, Z, X, Se, Sp, regime)
}
\arguments{
\item{enetgt.grid.out}{List object returned by the function \code{enetgt.grid}.}

\item{Z}{Group testing output from one of the functions \code{individual.assay.gen}, \code{masterpool.assay.gen}, \code{dorfman.assay.gen}, or \code{array.assay.gen}.}

\item{X}{Design matrix with first column a column of 1s.}

\item{Se}{A vector of testing sensitivities, where the first element is the
testing specificity for pools and the second entry is the 
test specificity for individual testing, if applicable.}

\item{Sp}{A vector of testing specificities, where the first element is the
testing specificity for pools and the second entry is the 
test specificity for individual testing, if applicable.}

\item{regime}{Testing regime which must be equal to one of \code{individual}, \code{masterpool}, \code{dorfman}, \code{array}, and \code{dorfman.individual}.}
}
\value{
A list object including the tuning parameter combinations selected by AIC, BIC, and ERIC.

Computes the AIC, BIC and ERIC criteria over a grid of lambda, theta pairs coming from \code{enetgt.grid.out}, 
which is an object returned by \code{enetgt.grid()}.
}
\description{
Chooses tuning parameters via AIC, BIC, and ERIC.
}
\examples{
data <- model0(200)
X <- data$X
Y.true <- data$Y
# subject individuals to individual testing
Se <- c(.94) # individual testing sensitivities
Sp <- c(.95) # individual testing specificities
assay.data <- individual.assay.gen(Y.true,Se,Sp,cj=1)
Z <- assay.data$Z
Y <- assay.data$Y
# compute the mle on the individual testing data:
mlegt.out <- mlegt(X,Y,Z,Se,Sp,delta=.01)
b.mle <- mlegt.out$b.mle
# compute adaptive elastic net estimator over a grid of tuning parameter values
n.lambda <- 8
n.theta <- 2
enetgt.grid.out <- enetgt.grid(X,Y,Z,Se,Sp,n.lambda,n.theta,weights = 1/abs(b.mle[-1]),delta=.01)
# make choices of the tuning parameters according to the aic, bic, and eric criteria
aic.bic.eric.enetgt.grid.out <- aic.bic.eric.enetgt.grid(enetgt.grid.out,Z,X,Se,Sp,"individual")
b.aenet.aic <- aic.bic.eric.enetgt.grid.out$b.enet.aic
b.aenet.bic <- aic.bic.eric.enetgt.grid.out$b.enet.bic
b.aenet.eric <- aic.bic.eric.enetgt.grid.out$b.enet.eric
b.aenet.aic 
b.aenet.bic
b.aenet.eric
}
