% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aenetgt.R
\name{enetgt.grid}
\alias{enetgt.grid}
\title{Computes the elastic net estimators with weighted l1 norm on group testing data over a grid of lambda and theta values.}
\usage{
enetgt.grid(X, Y, Z, Se, Sp, n.lambda = 5, n.theta = 3, weights = 1,
  delta = 0.001, E.approx = FALSE, verbose = FALSE, get.SEs = FALSE,
  ridge.include = FALSE)
}
\arguments{
\item{X}{Design matrix with first column a column of 1s.}

\item{Y}{Group testing output from one of the functions \code{individual.assay.gen}, \code{masterpool.assay.gen}, \code{dorfman.assay.gen}, or \code{array.assay.gen}.}

\item{Z}{Group testing output from one of the functions \code{individual.assay.gen}, \code{masterpool.assay.gen}, \code{dorfman.assay.gen}, or \code{array.assay.gen}.}

\item{Se}{A vector of testing sensitivities, where the first element is the
testing specificity for pools and the second entry is the 
test specificity for individual testing, if applicable.}

\item{Sp}{A vector of testing specificities, where the first element is the
testing specificity for pools and the second entry is the 
test specificity for individual testing, if applicable.}

\item{n.lambda}{Number of lambda values for which to compute the estimator.}

\item{n.theta}{Number of theta values for which to compute the estimator.}

\item{weights}{Vector of weights to be used in weighting the l1 penalty. Default is \code{weights=1}, which causes equal weights to be used for each coefficient.}

\item{delta}{Convergence criterion.}

\item{E.approx}{Logical.  If \code{TRUE} then E-step done with \code{EY.approx()}. If \code{FALSE}, then E-step done with \code{EY.exact()}.}

\item{verbose}{Logical. If \code{TRUE} then progress is reported after computation of the estimator at each tuning parameter pair.}

\item{get.SEs}{Logical.  If \code{TRUE} then estimated standard errors for the estimates are returned.}

\item{ridge.include}{Logical.  If \code{TRUE} then the sequence of theta values begins with 0 to include the ridge estimator.}
}
\value{
A list of output which includes an array \code{B.ENET} of solutions over the grid of lambda and theta values.

This function implements a penalized EM-algorithm to find the elastic net estimator with weighted l1 norm based on the observed data \code{X},
\code{Y}, \code{Z}, and the sensitivities and specificities in \code{Se}, \code{Sp} over a grid of lambda and theta values.  Only the size of 
the grid must be given, and the function chooses its own set of lambda and theta values.
}
\description{
Computes the elastic net estimators with weighted l1 norm on group testing data over a grid of lambda and theta values.
}
\examples{
# generate individual covariate values and disease statuses
N <- 160
data <- model1(N)
X <- data$X
Y.true <- data$Yi
Se <- c(.95) # set individual assay sensitivity
Sp <- c(.97) # set individual assay specificity
cj <- 1 # set size of master pools
# subject individuals to individual testing
assay.data <- individual.assay.gen(Y.true,Se,Sp,cj)
Z <- assay.data$Z
Y <- assay.data$Y
n.lambda <- 10
n.theta <- 2
# compute the elastic net estimator with weights = 1 over grid of lambda and theta values
enetgt.grid.out <- enetgt.grid(X, Y, Z, Se, Sp, n.lambda, n.theta, weights = 1) 
}
