% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aenetgt.R
\name{enetgt}
\alias{enetgt}
\title{Computes the elastic net estimator with weighted l1 norm on group testing data.}
\usage{
enetgt(X, Y, Z, Se, Sp, lambda, theta, weights = 1, binit = 1,
  delta = 0.001, E.approx = FALSE, get.SEs = FALSE)
}
\arguments{
\item{X}{Design matrix with first column a column of 1s.}

\item{Y}{Group testing output from one of the functions \code{individual.assay.gen}, \code{masterpool.assay.gen}, \code{dorfman.assay.gen}, or \code{array.assay.gen}.}

\item{Z}{Group testing output from one of the functions \code{individual.assay.gen}, \code{masterpool.assay.gen}, \code{dorfman.assay.gen}, or \code{array.assay.gen}.}

\item{Se}{A vector of testing sensitivities, where the first element is the
testing specificity for pools and the second entry is the 
test specificity for individual testing, if applicable.}

\item{Sp}{A vector of testing specificities, where the first element is the
testing specificity for pools and the second entry is the 
test specificity for individual testing, if applicable.}

\item{lambda}{Tuning parameter.}

\item{theta}{Ridge versus lasso penalty mixer.}

\item{weights}{Vector of weights to be used in weighting the l1 penalty. Default is \code{weights=1}, which causes equal weights to be used for each coefficient.}

\item{binit}{Initial values for EM-algorithm.}

\item{delta}{Convergence criterion.}

\item{E.approx}{Logical.  If \code{TRUE} then E-step done with \code{EY.approx()}. If \code{FALSE}, then E-step done with \code{EY.exact()}.}

\item{get.SEs}{Logical.  If \code{TRUE} then estimated standard errors for the estimates are returned.}
}
\value{
The elastic net estimator with weighted l1 norm under the choices of \code{lambda}, \code{theta}, and \code{weights}.

This function implements a penalized EM-algorithm to find the elastic net estimator with weighted l1 norm based on the observed data \code{X}, \code{Y}, \code{Z}, and the sensitivities and specificities in \code{Se}, \code{Sp}.
}
\description{
Computes the elastic net estimator with weighted l1 norm on group testing data.
}
\examples{
# generate individual covariate values and disease statuses
N <- 160
data <- model1(N)
X <- data$X
Y.true <- data$Yi
Se <- c(.95) # set individual assay sensitivity
Sp <- c(.97) # set individual assay specificity
cj <- 1 # set size of master pools
# subject individuals to individual testing
assay.data <- individual.assay.gen(Y.true,Se,Sp,cj)
Z <- assay.data$Z
Y <- assay.data$Y
# compute the elastic net estimator with weights = 1
enetgt.out <- enetgt(X, Y, Z, Se, Sp, lambda=10, theta=.5, weights = 1) 
# compute adaptive elastic net with weights from the elastic net estimator
a.enetgt.out <- enetgt(X, Y, Z, Se, Sp, lambda=.5, theta=.5, 
                       weights = 1/abs(enetgt.out$b.enet[-1])) 
}
