% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aenetgt.R
\name{EY.exact}
\alias{EY.exact}
\title{Computes conditional expectations of individual disease statuses.}
\usage{
EY.exact(Z, Y, X, b, Se, Sp)
}
\arguments{
\item{Z}{Group testing output from one of the functions \code{individual.assay.gen}, \code{masterpool.assay.gen}, \code{dorfman.assay.gen}, or \code{array.assay.gen}.}

\item{Y}{Group testing output from one of the functions \code{individual.assay.gen}, \code{masterpool.assay.gen}, \code{dorfman.assay.gen}, or \code{array.assay.gen}.}

\item{X}{Design matrix with first column a column of 1s.}

\item{b}{Parameter values at which to compute the conditional expectations.}

\item{Se}{A vector of testing sensitivities of length \code{max(Z[,3])}.}

\item{Sp}{A vector of testing specificities of length \code{max(Z[,3])}.}
}
\value{
The vector of conditional expectations.

This function computes the conditional expectations of each individual
disease status, conditional on the observed assay data and the disease 
statuses of all other individuals.  This function is used in the EM algorithm
performed by the functions \code{mlegt}, \code{enetgt}, \code{enetgt.grid}, and 
\code{enetgt.grid.0}.
}
\description{
Computes conditional expectations of individual disease statuses.
}
\examples{
# generate individual covariate values and disease statuses
N <- 100
data <- model1(N)
X <- data$X
Y.true <- data$Yi
Se <- c(.95,.92) # set master pool and individual assay sensitivity
Sp <- c(.97,.98) # set master pool and individual assay specificity
cj <- 4 # set size of master pools
# subject individuals to Dorfman testing
assay.data <- dorfman.assay.gen(Y.true,Se,Sp,cj)
Z <- assay.data$Z
Y <- assay.data$Y
b <- data$b
EY <- EY.exact(Z,Y,X,b,Se,Sp)
}
