% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EYexact}
\alias{EYexact}
\title{Computes conditional expectations of individual disease statuses for individual, master pool, or Dorfman testing}
\usage{
EYexact(Z, Y, eta, Se, Sp)
}
\arguments{
\item{Z}{Group testing output from one of the functions \code{individual.assay.gen}, \code{masterpool.assay.gen}, \code{dorfman.assay.gen}.}

\item{Y}{Group testing output from one of the functions \code{individual.assay.gen}, \code{masterpool.assay.gen}, \code{dorfman.assay.gen}.}

\item{eta}{the value of the linear predictor}

\item{Se}{A vector of testing sensitivities of length \code{max(Z[,3])}.}

\item{Sp}{A vector of testing specificities of length \code{max(Z[,3])}.}
}
\value{
The vector of conditional expectations.

This function computes the conditional expectations of each individual disease status given the observed assay data.
}
\description{
Computes conditional expectations of individual disease statuses for individual, master pool, or Dorfman testing
}
\examples{
# generate individual covariate values and disease statuses
N <- 100
data <- model1(N)
X <- data$X
Y.true <- data$Yi
Se <- c(.95,.92) # set master pool and individual assay sensitivity
Sp <- c(.97,.98) # set master pool and individual assay specificity
cj <- 4 # set size of master pools
# subject individuals to Dorfman testing
assay.data <- dorfman.assay.gen(Y.true,Se,Sp,cj)
Z <- assay.data$Z
Y <- assay.data$Y
b <- data$b
eta <- X \%*\% b
EY <- EYexact(Z,Y,eta,Se,Sp)
}
