% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aenetgt.R
\name{get.masterpool.cv.fold.data}
\alias{get.masterpool.cv.fold.data}
\title{Splits masterpool testing data into crossvalidation data sets}
\usage{
get.masterpool.cv.fold.data(X, Y, Z, K)
}
\arguments{
\item{X}{Design matrix with first column a column of 1s.}

\item{Y}{Group testing output from one of the functions \code{individual.assay.gen}, \code{masterpool.assay.gen}, \code{dorfman.assay.gen}, or \code{array.assay.gen}.}

\item{Z}{Group testing output from one of the functions \code{individual.assay.gen}, \code{masterpool.assay.gen}, \code{dorfman.assay.gen}, or \code{array.assay.gen}.}

\item{K}{The number of crossvalidation folds.}
}
\value{
List of length \code{K} of the crossvalidation training and testing data sets.
}
\description{
Splits masterpool testing data into crossvalidation data sets
}
\examples{
# generate individual covariate values and disease statuses
N <- 160
data <- model1(N)
X <- data$X
Y.true <- data$Yi
Se <- c(.95) # set master pool assay sensitivity
Sp <- c(.97) # set master pool assay specificity
cj <- 4 # set size of master pools
# subject individuals to master pool testing
assay.data <- masterpool.assay.gen(Y.true,Se,Sp,cj)
Z <- assay.data$Z
Y <- assay.data$Y
masterpool.cv.fold.data <- get.masterpool.cv.fold.data(X,Y,Z,K=5)
}
