% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logistic_enet}
\alias{logistic_enet}
\title{Compute the elastic net estimator for logistic regression}
\usage{
logistic_enet(Yr, Xr, lambda, gammar, theta, delta)
}
\arguments{
\item{Yr}{Response vector of 1s and 0s}

\item{Xr}{A design matrix with the first column a column of 1s}

\item{lambda}{The tuning parameter governing the strength of the elastic net penalty}

\item{gammar}{A vector of length \code{ncol(X) - 1} giving the weights applied to each covariate in the elastic net penalization}

\item{theta}{Value controlling the relative strength of the ridge and lasso penalties; 1 gives lasso.}

\item{delta}{Convergence tolerance}
}
\value{
a list with the estimated coefficients, etc.
}
\description{
Compute the elastic net estimator for logistic regression
}
