% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aenetgt.R
\name{logit}
\alias{logit}
\title{Compute probabilities based on the logit link.}
\usage{
logit(b, X)
}
\arguments{
\item{b}{Parameter vector.}

\item{X}{Design matrix with first column a column of 1s.}
}
\value{
The fitted probability according to the logit link function.
}
\description{
Compute probabilities based on the logit link.
}
