% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aenetgt.R
\name{mlegt}
\alias{mlegt}
\title{Computes the mle on group testing data.}
\usage{
mlegt(X, Y, Z, Se, Sp, binit = 1, delta = 0.001, E.approx = FALSE,
  get.SEs = FALSE)
}
\arguments{
\item{X}{Design matrix with first column a column of 1s.}

\item{Y}{Group testing output from one of the functions \code{individual.assay.gen}, \code{masterpool.assay.gen}, \code{dorfman.assay.gen}, or \code{array.assay.gen}.}

\item{Z}{Group testing output from one of the functions \code{individual.assay.gen}, \code{masterpool.assay.gen}, \code{dorfman.assay.gen}, or \code{array.assay.gen}.}

\item{Se}{A vector of testing sensitivities, where the first element is the
testing specificity for pools and the second entry is the 
test specificity for individual testing, if applicable.}

\item{Sp}{A vector of testing specificities, where the first element is the
testing specificity for pools and the second entry is the 
test specificity for individual testing, if applicable.}

\item{binit}{Parameter value at which to initialize the EM-algorithm. The default is \code{b=1}, for which an initial value is chosen internally.}

\item{delta}{Convergence criterion.}

\item{E.approx}{Logical.  If \code{TRUE} then E-step done with \code{EY.approx()}. If \code{FALSE}, then E-step done with \code{EY.exact()}.}

\item{get.SEs}{Logical.  If \code{TRUE} then estimated standard errors for the maximum likelihood estimators are returned.}
}
\value{
The maximum likelihood estimator.

This function implements an EM-algorithm to find the maximum likelihood estimator based on the observed data \code{X}, \code{Y}, \code{Z}, and the sensitivities and specificities in \code{Se}, \code{Sp}.
}
\description{
Computes the mle on group testing data.
}
\examples{
# generate individual covariate values and disease statuses
N <- 160
data <- model0(N)
X <- data$X
Y.true <- data$Yi
Se <- c(.95,.92) # set master pool and individual assay sensitivity
Sp <- c(.97,.98) # set master pool and individual assay specificity
cj <- 4 # set size of master pools
# subject individuals to dorfman testing
assay.data <- dorfman.assay.gen(Y.true,Se,Sp,cj)
Z <- assay.data$Z
Y <- assay.data$Y
mlegt.out <- mlegt(X, Y, Z, Se, Sp, delta = .01) # compute mle
}
