% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aenetgt.R
\name{model0}
\alias{model0}
\title{Generates data from model0.}
\usage{
model0(n)
}
\arguments{
\item{n}{Number of individuals for whom to generate data.}
}
\value{
List with design matrix \code{X}, disease statuses \code{Yi}, model identifier \code{model}, and parameter vector \code{b}.

This is a small model which is not really interesting but on which everything runs fast, because of its smallness.
}
\description{
Generates data from model0.
}
\examples{
N <- 100
data <- model0(N)
}
