% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aenetgt.R
\name{model3}
\alias{model3}
\title{Generates data from model3.}
\usage{
model3(n)
}
\arguments{
\item{n}{Number of individuals for whom to generate data.}
}
\value{
List with design matrix \code{X}, disease statuses \code{Yi}, model identifier \code{model}, and parameter vector \code{b}.
}
\description{
Generates data from model3.
}
\examples{
N <- 100
data <- model3(N)
}
